Feat: Gates para gestion de usuario

+++ app/Http/Controllers/Admin/UsersController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\User;
 use App\Role;
+use Gate;
 use Illuminate\Http\Request;
 
 
@@ -65,13 +66,16 @@ class UsersController extends Controller
      */
     public function edit(User $user)
     {
+        if(Gate::denies('edit-users')){
+            return redirect(route('admin.users.index'));
+        }
+
         $role = Role::all();
 
         return view('admin.users.edit')->with([
             'user' => $user,
             'roles' => $role,
         ]);
-        //dd($user);
     }
 
     /**
@@ -97,7 +101,10 @@ class UsersController extends Controller
      */
     public function destroy(User $user)
     {
-        //dd($user);
+        if(Gate::denies('delete-users')){
+            return redirect(route('admin.users.index'));
+        }
+
         $user->roles()->detach();
         $user->delete();
